Make Ref.apply() return trees usable in the largest scope possible
Previously for symbols contained in objects (prefixed by, let's say,
'pre'), we would return:
* an Ident if pre contained only static
object and packages;
* Select(This(moduleClassSymbol), sym) if a prefix contained a class.

However, this meant that in the second case, the generated tree would
require the macro to be expanded inside of the object, even though it
should be enough to just expand inside of the innermost class.
This was unexpected and confusing, so it was changed to not return
innermost module classes wrapped with This().

[Cherry-picked 238ba458b491c06e9f492673607157619b61d007]