Dealias before checking for member in lint (#22708)
Fixes #22705
Fixes #22706 
Fixes #22727 

Follow-up to https://github.com/scala/scala3/pull/22502 by inserting a
`dealias` when arriving at `target` type.

Refactored the body of `hidden` to make it easier to read. Adjusted the
doc for the same reason.

As a reminder to self, the original reason for special handling of
aliases was due to subclassing, but overrides are excluded. (One could
restore that warning for edge cases.)

The long doc explaining the handling of leading implicits is moved to
the end (as an appendix).

Despite best efforts, I was unable to make the doc longer than the code.
[Cherry-picked d36249284e79f33a897e3e53ea3967606f09c40c]