Show the Autofill completion case as what would be auto-filled (#22819)
Partial Fix for scalameta/metals#7274

We have in our completion provider an option to automatically fill a
case class constructor with named parameters:

```scala
case class A(x: Int, y: Int)

val a = A(@@)
          ^^ // would fill in here
val a = A(x = ???, y = ???)
          //  ^^^      ^^^  cursor marks added, so tabs can be used to move between the fields
```

However previously it was labeled `Autofill with default values`, which
would never pass editors' filters based on the
current state of input (most likely the name of a field).
The only way to find this option was to type something matching the
`Autofill with default values` label, which explains the
`auto` in the above issue.

We now provide the actual provided fields as the label, so the above
would show `x = ???, y = ???` instead, which is both
more intuitive for the user, and more friendly to editors.
[Cherry-picked 744ba922a1729e7670a0db6de981932e60b2b56e]