Dealias for unused param check (#23256)
Fixes #23250 

The check for "unused implicit parameter" considers the type of the
parameter, but it should dealias first.

In addition, take an abstract type as not warnable, unless it has an
upper bound that is warnable (has non-universal members).

Note that `isSingleton` dealiases.

There was an ancient (OG) typo of `&` for `&&` which has gone unfixed
despite other edits. The right operand was always just an instanceof.
Not sure if the previous idiom saves an opcode, but now it looks like
other code nearby.
[Cherry-picked 2d1865510867aba338813e0ac265f6a8fa70ed35]