Constrain by deepening results in more implicit searches
We used to look at deep expected type when the result of an implicit was ambiguous.
This could add more constraints which could resolve the ambiguity.

We now do the same also if the search type has uninstantiated type variables. In that
case, consulting more context might further constrain type variables, which might in
turn enlarge the implicit scope so that a solution can be found.

Fixes #23526

[Cherry-picked 45f7ef6e5d1f9a3b45b77b933470de43760b0c72]