Ignore selection prototypes when typing type applications
Ignore selection prototypes at first when typing type applications.
If we need them later for overloading disambiguation, reveal the ignored
type.

The reason for doing this is that a selection might come from an extension method,
and in this case we should not require the selected name as a member of the result.

This change breaks one test (overloading-specifity-2.scala) that explicitly
tested that we don't consult implicit arguments for disambiguation since the
expected type was a selection that already determined the outcome. This is logic
no longer holds. We have to see whether this change breaks any code in practice.

Fixes #23773

[Cherry-picked 30aa80596aa8f0b75fd9daa8ff15511902ca18bc][modified]